refactor(routes): re-write loadRoute to helpers directory
