+++ routes/api.php
@@ -3,6 +3,8 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
+require_once app_path('Helpers/RouteHelper.php');
+
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -18,19 +20,5 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-function loadRoutesApi($dir)
-{
-    if (!is_dir($dir)) return;
 
-    $files = scandir($dir);
-    foreach ($files as $file) {
-        if ($file == '.' || $file == '..') continue;
-
-        $filePath = $dir . '/' . $file;
-        if (is_file($filePath) && pathinfo($filePath)['extension'] === 'php')
-            require_once $filePath;
-        else if (is_dir($filePath)) loadRoutes($filePath);
-    }
-}
-
-loadRoutesApi(__DIR__ . '/Handler/api');
+loadRoutes(__DIR__ . '/Handler/api');
